Refactor: AuthorizationComponent 클래스의 위치 변경
