+++ src/main/java/podo/odeego/web/security/AuthorizationComponent.java
@@ -1,4 +1,4 @@
-package podo.odeego.web.security.service;
+package podo.odeego.web.security;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
